requirements.txt
requests>=2.0
python-docx>=0.8.11
beautifulsoup4>=4.12.2
openai>=0.27.0
langchain-openai>=0.0.6
